#114 - Hybrid flow : add ability to set response_type directly during authorization url request
