+++ oxd-server/src/main/java/org/gluu/oxd/server/op/BaseOperation.java
@@ -94,10 +94,6 @@ public abstract class BaseOperation<T extends IParams> implements IOperation<T> 
         return getInstance(ConfigurationService.class);
     }
 
-    public UpdateRegisteredClientService getUpdateRegisteredClientService() {
-        return getInstance(UpdateRegisteredClientService.class);
-    }
-
     public OxAuthCryptoProvider getCryptoProvider() throws Exception {
         OxdServerConfiguration conf = getConfigurationService().get();
         return new OxAuthCryptoProvider(conf.getCryptProviderKeyStorePath(), conf.getCryptProviderKeyStorePassword(), conf.getCryptProviderDnName());
